\name{abundance.surv}
\alias{abundance.surv}
\title{Alternative abundance estimators for open population models}
\usage{
  abundance.surv(x, model)
}
\arguments{
  \item{x}{dataframe with a field \code{ch} which is the
  capture history string of 0s & 1s}

  \item{model}{fitted model}
}
\value{
  List of results
}
\description{
  Alternative abundance estimators for open population
  models
}
\author{
  Jeff Laake
}

