% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noCalf.r
\name{noCalf_Abundance}
\alias{noCalf_Abundance}
\title{Abundance estimation (JS1) without cavles}
\usage{
noCalf_Abundance(er, popan.results, chat, delta = Inf)
}
\arguments{
\item{er}{dataframe with a field \code{ch} which is the capture history
string of 0s & 1s}

\item{popan.results}{list of fitted models}

\item{chat}{Overdispersion value}

\item{delta}{value of delta aicc to use for model averaging set. Default is Inf to use all.}

\item{alternate}{If TRUE do not median center MT and use JS1 approach for abundance estimation}
}
\value{
List of abundance estimates
}
\description{
Abundance estimation (JS1) without cavles
}
\author{
Jeff Laake
}
