% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.r
\name{abundance.surv}
\alias{abundance.surv}
\title{Alternative abundance estimators for open population models}
\usage{
abundance.surv(x, model, minyear)
}
\arguments{
\item{x}{dataframe with a field \code{ch} which is the capture history
string of 0s & 1s}

\item{model}{fitted model}

\item{minyear}{first year of data}

\item{Calf}{if FALSE they are excluded in the estimate in the year first seen}
}
\value{
List of results
}
\description{
Alternative abundance estimators for open population models
}
\author{
Jeff Laake
}
