\name{summary.stats.plots}
\alias{summary.stats.plots}
\title{Produces a set of summary statistics and plots for a subset of the PCFA data}
\usage{
  summary.stats.plots(x, ch)
}
\arguments{
  \item{x}{dataframe (or subset) of \code{\link{PCFA}}}

  \item{ch}{capture history dataframe created by
  \code{\link{create.chmat}} using same data}
}
\value{
  None
}
\description{
  Produces a set of summary statistics and plots for a
  subset of the PCFA data
}
\author{
  Jeff Laake
}
\seealso{
  \code{\link{movement.table}}
}

